replace react dispatch with callbacks
