+++ packages/sdk/src/chains/Substrate/index.ts
@@ -1,1 +1,12 @@
-export * from './utils';
+export {
+  substrateSocketConnect,
+  getAssetBalance,
+  getBasicFee,
+  getNativeTokenBalance,
+  loadAccounts,
+  retrieveChainInfo,
+  deposit,
+  throwErrorIfAny,
+  handleTxExtrinsicResult,
+  calculateBigNumber,
+} from './utils';
