+++ packages/sdk/src/chains/Substrate/utils/index.ts
@@ -1,0 +1,13 @@
+export { substrateSocketConnect } from './substrateSocketConnect';
+export { getAssetBalance } from './getAssetBalance';
+export { getBasicFee } from './getBasicFee';
+export { getNativeTokenBalance } from './getNativeTokenBalance';
+
+export { loadAccounts } from './loadAccounts';
+export { retrieveChainInfo } from './retrieveChainInfo';
+export {
+  deposit,
+  throwErrorIfAny,
+  handleTxExtrinsicResult,
+  calculateBigNumber,
+} from './depositFns';
